AshOban (ash_oban v0.2.4)

Tools for working with AshOban triggers.

Types

result()

@type result() :: %{
  discard: non_neg_integer(),
  cancelled: non_neg_integer(),
  success: non_neg_integer(),
  failure: non_neg_integer(),
  snoozed: non_neg_integer(),
  queues_not_drained: [atom()]
}

triggerable()

@type triggerable() ::
  Ash.Resource.t() | {Ash.Resource.t(), atom()} | Ash.Domain.t() | atom()

Functions

authorize?()

@spec authorize?() :: boolean()

build_trigger(record, trigger, opts \\ [])

Builds a specific trigger for the record provided, but does not insert it into the database.

Options

:actor - the actor to set on the job. Requires configuring an actor persister.

All other options are passed through to Oban.Worker.new/2

config(domains, base, opts \\ [])

Alters your oban configuration to include the required AshOban configuration.

Options

:require? (boolean/0) - Whether to require queues and plugins to be defined in your oban config. This can be helpful to allow the ability to split queues between nodes. See https://hexdocs.pm/oban/splitting-queues.html The default value is true.

do_schedule_and_run_triggers(resources_or_domains_or_otp_apps, opts)

lookup_actor(actor_json)

@spec lookup_actor(actor_json :: any()) :: any()

run_trigger(record, trigger, opts \\ [])

Runs a specific trigger for the record provided.

Options

:actor - the actor to set on the job. Requires configuring an actor persister.

All other options are passed through to Oban.Worker.new/2

schedule(resource, trigger, opts \\ [])

Schedules all relevant jobs for the provided trigger or scheduled action

Options

:actor - the actor to set on the job. Requires configuring an actor persister.

schedule_and_run_triggers(resources_or_domains_or_otp_apps, opts \\ [])

@spec schedule_and_run_triggers(
  triggerable() | [triggerable()],
  keyword()
) :: result()

Runs the schedulers for the given resource, domain, or otp_app, or list of resources, domains, or otp_apps.

Options:

drain_queues? - Defaults to false, drains the queues after scheduling. This is primarily for testing

queue, with_limit, with_recursion, with_safety, with_scheduled - passed through to Oban.drain_queue/2, if it is called

scheduled_actions? - Defaults to false, unless a scheduled action name was explicitly provided. Schedules all applicable scheduled actions.

triggers? - Defaults to true, schedules all applicable scheduled actions.

actor - The actor to schedule and run the triggers with

oban - The oban module to use. Defaults to Oban

If the input is:

a list - each item is passed into schedule_and_run_triggers/1, and the results are merged together.

an otp_app - each domain configured in the ash_domains of that otp_app is passed into schedule_and_run_triggers/1, and the results are merged together.

a domain - each reosurce configured in that domain is passed into schedule_and_run_triggers/1, and the results are merged together.

a tuple of {resource, :trigger_name} - that trigger is scheduled, and the results are merged together.

a resource - each trigger configured in that resource is scheduled, and the results are merged together.

stacktrace(arg1)

store_actor(args, actor)

@spec store_actor(args :: map(), actor :: any()) :: any()